% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataSets.R
\name{attachHoubaBigMatrix}
\alias{attachHoubaBigMatrix}
\title{Attach a houba file and return the bigmemory::big.matrix}
\usage{
attachHoubaBigMatrix(houba_file)
}
\arguments{
\item{houba_file}{Path to houba file (without .desc extension)}
}
\value{
bigmemory::big.matrix object
}
\description{
Attach a houba file and return the bigmemory::big.matrix
}
