% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataSets.R
\name{DatasetParquetUrlList}
\alias{DatasetParquetUrlList}
\title{List of the huggingface datasets for the paper
"Incorporating LLM Embeddings for Variation Across the Human Genome"}
\usage{
DatasetParquetUrlList()
}
\value{
A character vector of dataset names
}
\description{
List of the huggingface datasets for the paper
"Incorporating LLM Embeddings for Variation Across the Human Genome"
}
