% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataSets.R
\name{embeddingSummary}
\alias{embeddingSummary}
\title{Quick summary for houba mmatrix}
\usage{
embeddingSummary(embMat)
}
\arguments{
\item{embMat}{houba mmatrix}
}
\value{
List with dim, colMeans, rowMeans
}
\description{
Quick summary for houba mmatrix
}
