% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataSets.R
\name{getPcaScores}
\alias{getPcaScores}
\title{Get PCA scores from houbaPCA result}
\usage{
getPcaScores(houbaPCA_res)
}
\arguments{
\item{houbaPCA_res}{List returned by houbaPCA (with 'pca' and 'houbaM')}
}
\value{
Matrix of principal component scores (variants x PCs)
}
\description{
Get PCA scores from houbaPCA result
}
