% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataSets.R
\name{houbaPCA}
\alias{houbaPCA}
\title{PCA using bigPCAcpp on houba mmatrix or bigmemory::big.matrix}
\usage{
houbaPCA(embMat, center = TRUE, scale = TRUE, ncomp = 10)
}
\arguments{
\item{embMat}{houba mmatrix path or bigmemory::big.matrix}

\item{center}{logical, whether to center columns}

\item{scale}{logical, whether to scale columns}

\item{ncomp}{number of principal components to compute}
}
\value{
PCA result object from bigPCAcpp
}
\description{
PCA using bigPCAcpp on houba mmatrix or bigmemory::big.matrix
}
