% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataSets.R
\name{infoSummary}
\alias{infoSummary}
\title{Quick summary for houba info mmatrix}
\usage{
infoSummary(infoMat)
}
\arguments{
\item{infoMat}{houba mmatrix}
}
\value{
List with dim, unique chroms, and example rsids
}
\description{
Quick summary for houba info mmatrix
}
