% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataSets.R
\name{IterateEmbeddingsMatrixBatches}
\alias{IterateEmbeddingsMatrixBatches}
\title{Iterate over embeddings as matrix batches from a local DuckDB file}
\usage{
IterateEmbeddingsMatrixBatches(
  chunk_size = 1e+05,
  db_path = "local_embeddings.duckdb",
  table_name = "embeddings"
)
}
\arguments{
\item{chunk_size}{Number of rows per batch}

\item{db_path}{Path to the local DuckDB database file}

\item{table_name}{Name of the table in the DuckDB database}
}
\value{
A list of embedding batches
}
\description{
Iterate over embeddings as matrix batches from a local DuckDB file
}
