% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataSets.R
\name{plotPcaDims}
\alias{plotPcaDims}
\title{Plot PCA dimensions using ggplot2, colored by annotation}
\usage{
plotPcaDims(pc_scores, info_df, annotation_col = "chrom", dim1 = 1, dim2 = 2)
}
\arguments{
\item{pc_scores}{Matrix of principal component scores (variants x PCs)}

\item{info_df}{Data frame with variant info (must contain annotation column)}

\item{annotation_col}{Name of column in info_df to color by (e.g. 'gwas')}

\item{dim1}{First PC dimension to plot (default: 1)}

\item{dim2}{Second PC dimension to plot (default: 2)}
}
\description{
Plot PCA dimensions using ggplot2, colored by annotation
}
