% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataSets.R
\name{plotPCSpatialCorrelation}
\alias{plotPCSpatialCorrelation}
\title{Plot spatial correlation between PC scores and genomic position, faceted by chromosome}
\usage{
plotPCSpatialCorrelation(pc_scores, info_df, pc = 1)
}
\arguments{
\item{pc_scores}{Matrix of principal component scores (variants x PCs)}

\item{info_df}{Data frame with variant info (must contain 'chrom' and 'pos')}

\item{pc}{Which principal component to plot (default: 1)}
}
\description{
Plot spatial correlation between PC scores and genomic position, faceted by chromosome
}
