% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataSets.R
\name{writeEmbeddingsHoubaFromDuckDB}
\alias{writeEmbeddingsHoubaFromDuckDB}
\title{Write embeddings to houba mmatrix and return info as data.frame from a local DuckDB file}
\usage{
writeEmbeddingsHoubaFromDuckDB(
  dbPath = "local_embeddings.duckdb",
  tableName = "embeddings",
  embeddingCol = "embedding",
  batchSize = 1e+05,
  embeddingDim = 3072,
  embeddingFile = gsub("\\\\.duckdb$", ".houba", dbPath)
)
}
\arguments{
\item{dbPath}{Path to the local DuckDB database file}

\item{tableName}{Name of the table in the DuckDB database}

\item{embeddingCol}{Name of the embeddings column}

\item{batchSize}{Number of rows per batch}

\item{embeddingDim}{Dimension of each embedding vector}

\item{embeddingFile}{Path for houba mmatrix file}
}
\value{
An object of class 'HoubaEmbeddings' with mmatrix, info data.frame, and houba file path
}
\description{
Write embeddings to houba mmatrix and return info as data.frame from a local DuckDB file
}
