% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataSets.R
\name{CopyParquetToDuckDB}
\alias{CopyParquetToDuckDB}
\title{Copy remote parquet files into a local DuckDB database file}
\usage{
CopyParquetToDuckDB(
  db_path = "local_embeddings.duckdb",
  urlList = DatasetParquetUrlList(),
  table_name = "embeddings"
)
}
\arguments{
\item{db_path}{Path to the local DuckDB database file}

\item{urlList}{Character vector of parquet URLs}

\item{table_name}{Name of the table to create in the DuckDB database}

\item{overwrite}{Whether to overwrite the existing database}
}
\description{
Copy remote parquet files into a local DuckDB database file
}
