% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDataSets.R
\name{houbaPCA}
\alias{houbaPCA}
\title{PCA using bigPCAcpp on houba mmatrix or bigmemory::big.matrix}
\usage{
houbaPCA(
  embMat = "local_embeddings.houba",
  center = TRUE,
  scale = TRUE,
  ncomp = 15
)
}
\arguments{
\item{embMat}{houba mmatrix path or bigmemory::big.matrix}

\item{center}{logical, whether to center columns}

\item{scale}{logical, whether to scale columns}

\item{ncomp}{number of principal components to compute}
}
\value{
List with PCA result object and houbaM big.matrix
}
\description{
PCA using bigPCAcpp on houba mmatrix or bigmemory::big.matrix
}
